+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+* fixed mistakes in `module` Python documentation
 
 ## [3.1.7] - 2020-11-26 14:30:00+02:00 (urgency: medium)
 * **WARNING:** the function signature of `netlist_utilities::get_subgraph_function` has changed